update formats and something else

+++ include/value.h
@@ -40,8 +40,8 @@ namespace mysql {
 
  @return The size in bytes of a particular field
 */
-int calc_field_size(unsigned char column_type, const unsigned char *field_ptr,
-                    uint16_t metadata);
+uint32_t calc_field_size(unsigned char column_type, const unsigned char *field_ptr,
+                    uint32_t metadata);
 
 uint8_t calc_newdecimal_size(uint8_t m, uint8_t d);
 
@@ -54,12 +54,12 @@ uint8_t calc_newdecimal_size(uint8_t m, uint8_t d);
 class Value
 {
 public:
-    Value(enum system::enum_field_types type, uint16_t metadata, const char *storage) :
+    Value(enum system::enum_field_types type, uint32_t metadata, const char *storage) :
       m_type(type), m_storage(storage), m_metadata(metadata), m_is_null(false)
     {
-      m_size = calc_field_size((unsigned char)type,
-                              (const unsigned char*)storage,
-                              metadata);
+      m_size = calc_field_size((unsigned char) type,
+                               (const unsigned char*) storage,
+                               metadata);
       // metadata: 16bits, lower byte is real type, high byte is length
       switch(type) {
           case mysql::system::MYSQL_TYPE_SET:
